data-pipeline/app/requirements.txt
pybaseball==2.2.1
pandas==1.4.2
google-cloud-storage==2.3.0
gcsfs==2022.3.0
python-dotenv==0.20.0